create regeneration button command
